Attempt to improve reliabiliy and speed of tests
git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@104142 7033f55c-6c29-4ad8-afba-765cb616e738

+++ rm-automation/src/main/java/org/alfresco/po/common/util/Utils.java
@@ -80,7 +80,7 @@ public final class Utils implements ApplicationContextAware
      *
      * @return
      */
-    public static WebDriverWait webDriverWait(int waitSeconds)
+    public static WebDriverWait webDriverWait(long waitSeconds)
     {
         return new WebDriverWait(getWebDriver(), waitSeconds);
     }
